Updating to latest FlutterFlow output.

+++ lib/backend/supabase/database/tables/pedidos_cliente.dart
@@ -36,6 +36,6 @@ class PedidosClienteRow extends SupabaseDataRow {
   String? get pedido => getField<String>('pedido');
   set pedido(String? value) => setField<String>('pedido', value);
 
-  int? get quanty => getField<int>('quanty');
-  set quanty(int? value) => setField<int>('quanty', value);
+  double? get quanty => getField<double>('quanty');
+  set quanty(double? value) => setField<double>('quanty', value);
 }
